Removed support for multi-gpu training.
